+++ packages/plugin-source-build/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rsbuild/plugin-source-build
 
+## 0.3.4
+
+### Patch Changes
+
+- Updated dependencies [3ebf3b5]
+  - @rsbuild/shared@0.3.4
+  - @rsbuild/core@0.3.4
+  - @rsbuild/monorepo-utils@0.3.4
+
 ## 0.3.3
 
 ### Patch Changes